Speedup test build by compile the fake server just once
