+++ test/CMakeLists.txt
@@ -14,6 +14,10 @@ include_directories(${CMAKE_SOURCE_DIR}/src
 
 include(nextcloud_add_test.cmake)
 
+set(CMAKE_AUTOMOC TRUE)
+add_library(syncenginetestutils STATIC syncenginetestutils.cpp)
+target_link_libraries(syncenginetestutils PUBLIC ${APPLICATION_EXECUTABLE}sync Qt5::Test)
+
 nextcloud_add_test(NextcloudPropagator "")
 
 IF(BUILD_UPDATER)
@@ -45,31 +49,31 @@ nextcloud_add_test(ClientSideEncryption "")
 nextcloud_add_test(ExcludedFiles "")
 
 nextcloud_add_test(Utility "")
-nextcloud_add_test(SyncEngine "syncenginetestutils.h")
-nextcloud_add_test(SyncVirtualFiles "syncenginetestutils.h")
-nextcloud_add_test(SyncMove "syncenginetestutils.h")
-nextcloud_add_test(SyncDelete "syncenginetestutils.h")
-nextcloud_add_test(SyncConflict "syncenginetestutils.h")
-nextcloud_add_test(SyncFileStatusTracker "syncenginetestutils.h")
-nextcloud_add_test(Download "syncenginetestutils.h")
-nextcloud_add_test(ChunkingNg "syncenginetestutils.h")
-nextcloud_add_test(AsyncOp "syncenginetestutils.h")
-nextcloud_add_test(UploadReset "syncenginetestutils.h")
-nextcloud_add_test(AllFilesDeleted "syncenginetestutils.h")
-nextcloud_add_test(Blacklist "syncenginetestutils.h")
-nextcloud_add_test(LocalDiscovery "syncenginetestutils.h")
-nextcloud_add_test(RemoteDiscovery "syncenginetestutils.h")
-nextcloud_add_test(Permissions "syncenginetestutils.h")
-nextcloud_add_test(SelectiveSync "syncenginetestutils.h")
-nextcloud_add_test(DatabaseError "syncenginetestutils.h")
-nextcloud_add_test(LockedFiles "syncenginetestutils.h;../src/gui/lockwatcher.cpp")
+nextcloud_add_test(SyncEngine "")
+nextcloud_add_test(SyncVirtualFiles "")
+nextcloud_add_test(SyncMove "")
+nextcloud_add_test(SyncDelete "")
+nextcloud_add_test(SyncConflict "")
+nextcloud_add_test(SyncFileStatusTracker "")
+nextcloud_add_test(Download "")
+nextcloud_add_test(ChunkingNg "")
+nextcloud_add_test(AsyncOp "")
+nextcloud_add_test(UploadReset "")
+nextcloud_add_test(AllFilesDeleted "")
+nextcloud_add_test(Blacklist "")
+nextcloud_add_test(LocalDiscovery "")
+nextcloud_add_test(RemoteDiscovery "")
+nextcloud_add_test(Permissions "")
+nextcloud_add_test(SelectiveSync "")
+nextcloud_add_test(DatabaseError "")
+nextcloud_add_test(LockedFiles "../src/gui/lockwatcher.cpp")
 nextcloud_add_test(FolderWatcher "${FolderWatcher_SRC}")
 
 if( UNIX AND NOT APPLE )
     nextcloud_add_test(InotifyWatcher "${FolderWatcher_SRC}")
 endif(UNIX AND NOT APPLE)
 
-nextcloud_add_benchmark(LargeSync "syncenginetestutils.h")
+nextcloud_add_benchmark(LargeSync "")
 
 SET(FolderMan_SRC ../src/gui/folderman.cpp)
 list(APPEND FolderMan_SRC ../src/gui/folder.cpp )
@@ -106,7 +110,7 @@ list(APPEND RemoteWipe_SRC ${RemoteWipe_SRC})
 list(APPEND RemoteWipe_SRC stubremotewipe.cpp )
 nextcloud_add_test(RemoteWipe "${RemoteWipe_SRC}")
 
-nextcloud_add_test(OAuth "syncenginetestutils.h;../src/gui/creds/oauth.cpp")
+nextcloud_add_test(OAuth "../src/gui/creds/oauth.cpp")
 
 configure_file(test_journal.db "${PROJECT_BINARY_DIR}/bin/test_journal.db" COPYONLY)
