+++ test/testchunkingng.cpp
@@ -275,7 +275,7 @@ private slots:
             QCOMPARE(items[0]->_file, QLatin1String("A"));
             SyncJournalFileRecord record;
             QVERIFY(fakeFolder.syncJournal().getFileRecord(QByteArray("A/a0"), &record));
-            QCOMPARE(record._etag, fakeFolder.remoteModifier().find("A/a0")->etag.toUtf8());
+            QCOMPARE(record._etag, fakeFolder.remoteModifier().find("A/a0")->etag);
         };
         auto connection = connect(&fakeFolder.syncEngine(), &SyncEngine::aboutToPropagate, checkEtagUpdated);
         QVERIFY(fakeFolder.syncOnce());
